+++ src/Oro/Bundle/RedirectBundle/Routing/SluggableUrlGenerator.php
@@ -2,12 +2,12 @@
 
 namespace Oro\Bundle\RedirectBundle\Routing;
 
-use Oro\Bundle\FrontendLocalizationBundle\Manager\UserLocalizationManager;
-use Oro\Bundle\RedirectBundle\Cache\UrlDataStorage;
-use Oro\Bundle\RedirectBundle\Cache\UrlStorageCache;
-use Oro\Bundle\RedirectBundle\Provider\ContextUrlProviderRegistry;
 use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
 use Symfony\Component\Routing\RequestContext;
+
+use Oro\Bundle\FrontendLocalizationBundle\Manager\UserLocalizationManager;
+use Oro\Bundle\RedirectBundle\Provider\SluggableUrlProviderInterface;
+use Oro\Bundle\RedirectBundle\Provider\ContextUrlProviderRegistry;
 
 class SluggableUrlGenerator implements UrlGeneratorInterface
 {
@@ -21,14 +21,14 @@ class SluggableUrlGenerator implements UrlGeneratorInterface
     private $generator;
 
     /**
-     * @var UrlStorageCache
-     */
-    private $cache;
-
-    /**
      * @var ContextUrlProviderRegistry
      */
     private $contextUrlProvider;
+
+    /**
+     * @var SluggableUrlProviderInterface
+     */
+    private $sluggableUrlProvider;
 
     /**
      * @var UserLocalizationManager
@@ -36,16 +36,16 @@ class SluggableUrlGenerator implements UrlGeneratorInterface
     private $userLocalizationManager;
 
     /**
-     * @param UrlStorageCache $cache
+     * @param SluggableUrlProviderInterface $sluggableUrlProvider
      * @param ContextUrlProviderRegistry $contextUrlProvider
      * @param UserLocalizationManager $userLocalizationManager
      */
     public function __construct(
-        UrlStorageCache $cache,
+        SluggableUrlProviderInterface $sluggableUrlProvider,
         ContextUrlProviderRegistry $contextUrlProvider,
         UserLocalizationManager $userLocalizationManager
     ) {
-        $this->cache = $cache;
+        $this->sluggableUrlProvider = $sluggableUrlProvider;
         $this->contextUrlProvider = $contextUrlProvider;
         $this->userLocalizationManager = $userLocalizationManager;
     }
@@ -70,25 +70,13 @@ class SluggableUrlGenerator implements UrlGeneratorInterface
     private function generateSluggableUrl($name, $parameters)
     {
         $contextUrl = $this->getContextUrl($parameters);
+        $localizationId = $this->getLocalizationId();
 
         $url = null;
-        $localization = $this->userLocalizationManager->getCurrentLocalization();
-        $localizationId = null;
-        if ($localization) {
-            $localizationId = $localization->getId();
-        }
-        $urlDataStorage = $this->getUrlDataStorage($name, $parameters);
-        if ($urlDataStorage) {
-            // For context aware URLs slug may be used as item part
-            if ($contextUrl && $slug = $urlDataStorage->getSlug($parameters, $localizationId)) {
-                $url = $slug;
-            }
 
-            // For URLs without context only full URL is acceptable
-            if (!$url) {
-                $url = $urlDataStorage->getUrl($parameters, $localizationId);
-            }
-        }
+        $this->sluggableUrlProvider->setContextUrl($contextUrl);
+
+        $url = $this->sluggableUrlProvider->getUrl($name, $parameters, $localizationId);
 
         // If no Slug based URL is available - generate URL with base generator logic
         if (!$url) {
@@ -120,16 +108,6 @@ class SluggableUrlGenerator implements UrlGeneratorInterface
     public function setBaseGenerator(UrlGeneratorInterface $generator)
     {
         $this->generator = $generator;
-    }
-
-    /**
-     * @param string $name
-     * @param array $parameters
-     * @return null|UrlDataStorage
-     */
-    private function getUrlDataStorage($name, $parameters)
-    {
-        return $this->cache->getUrlDataStorage($name, $parameters);
     }
 
     /**
@@ -173,5 +151,14 @@ class SluggableUrlGenerator implements UrlGeneratorInterface
         $urlParts[] = ltrim($url, '/');
 
         return '/' . implode('/', $urlParts);
+    }
+
+    /**
+     * @return null|int
+     */
+    private function getLocalizationId()
+    {
+        $localization = $this->userLocalizationManager->getCurrentLocalization();
+        return $localization ? $localization->getId() : null;
     }
 }