+++ src/Oro/Bundle/ShippingBundle/Context/LineItem/Collection/Doctrine/Factory/DoctrineShippingLineItemCollectionFactory.php
@@ -4,6 +4,7 @@ namespace Oro\Bundle\ShippingBundle\Context\LineItem\Collection\Doctrine\Factory
 
 use Oro\Bundle\ShippingBundle\Context\LineItem\Collection\Doctrine\DoctrineShippingLineItemCollection;
 use Oro\Bundle\ShippingBundle\Context\LineItem\Collection\Factory\ShippingLineItemCollectionFactoryInterface;
+use Oro\Bundle\ShippingBundle\Context\LineItem\Collection\ShippingLineItemCollectionInterface;
 use Oro\Bundle\ShippingBundle\Context\ShippingLineItemInterface;
 
 class DoctrineShippingLineItemCollectionFactory implements ShippingLineItemCollectionFactoryInterface
@@ -11,7 +12,7 @@ class DoctrineShippingLineItemCollectionFactory implements ShippingLineItemColle
     /**
      * {@inheritDoc}
      */
-    public function createShippingLineItemCollection(array $shippingLineItems)
+    public function createShippingLineItemCollection(array $shippingLineItems): ShippingLineItemCollectionInterface
     {
         foreach ($shippingLineItems as $shippingLineItem) {
             if (!$shippingLineItem instanceof ShippingLineItemInterface) {